HHH-20123 Address the split package issue in hibernate-micrometer and hibernate-core
