+++ hibernate-micrometer/src/main/java/org/hibernate/stat/HibernateQueryMetrics.java
@@ -4,19 +4,11 @@
  */
 package org.hibernate.stat;
 
-import io.micrometer.common.lang.NonNullApi;
-import io.micrometer.common.lang.NonNullFields;
-import io.micrometer.core.instrument.*;
+import io.micrometer.core.instrument.MeterRegistry;
+import io.micrometer.core.instrument.Tag;
+import io.micrometer.core.instrument.Tags;
 import io.micrometer.core.instrument.binder.MeterBinder;
-
 import org.hibernate.SessionFactory;
-import org.hibernate.engine.spi.SessionFactoryImplementor;
-import org.hibernate.event.service.spi.EventListenerRegistry;
-import org.hibernate.event.spi.EventType;
-import org.hibernate.event.spi.PostLoadEvent;
-import org.hibernate.event.spi.PostLoadEventListener;
-
-import java.util.concurrent.TimeUnit;
 
 /**
  * A {@link MeterBinder} implementation that provides Hibernate query metrics. It exposes the
@@ -25,16 +17,11 @@ import java.util.concurrent.TimeUnit;
  * <p>
  * Be aware of the potential for high cardinality of unique Hibernate queries executed by your
  * application when considering using this {@link MeterBinder}.
+ *
+ * @deprecated Use {@link org.hibernate.orm.micrometer.HibernateQueryMetrics} instead.
  */
-@NonNullApi
-@NonNullFields
-public class HibernateQueryMetrics implements MeterBinder {
-
-	private static final String SESSION_FACTORY_TAG_NAME = "entityManagerFactory";
-
-	private final Iterable<Tag> tags;
-
-	private final SessionFactory sessionFactory;
+@Deprecated(since = "7.3", forRemoval = true)
+public class HibernateQueryMetrics extends org.hibernate.orm.micrometer.HibernateQueryMetrics {
 
 	/**
 	 * Create {@code HibernateQueryMetrics} and bind to the specified meter registry.
@@ -76,110 +63,6 @@ public class HibernateQueryMetrics implements MeterBinder {
 	 * @param tags additional tags
 	 */
 	public HibernateQueryMetrics(SessionFactory sessionFactory, String sessionFactoryName, Iterable<Tag> tags) {
-		this.tags = Tags.concat( tags, SESSION_FACTORY_TAG_NAME, sessionFactoryName );
-		this.sessionFactory = sessionFactory;
-	}
-
-	@Override
-	public void bindTo(MeterRegistry meterRegistry) {
-		if ( sessionFactory instanceof SessionFactoryImplementor ) {
-			EventListenerRegistry eventListenerRegistry = ( (SessionFactoryImplementor) sessionFactory ).getEventEngine().getListenerRegistry();
-			MetricsEventHandler metricsEventHandler = new MetricsEventHandler( meterRegistry );
-			eventListenerRegistry.appendListeners( EventType.POST_LOAD, metricsEventHandler );
-		}
-	}
-
-	class MetricsEventHandler implements PostLoadEventListener {
-
-		private final MeterRegistry meterRegistry;
-
-		MetricsEventHandler(MeterRegistry meterRegistry) {
-			this.meterRegistry = meterRegistry;
-		}
-
-		@Override
-		public void onPostLoad(PostLoadEvent event) {
-			registerQueryMetric( event.getFactory().getStatistics() );
-		}
-
-		void registerQueryMetric(Statistics statistics) {
-			for ( String query : statistics.getQueries() ) {
-				QueryStatistics queryStatistics = statistics.getQueryStatistics( query );
-
-				FunctionCounter.builder(
-						"hibernate.query.cache.requests",
-						queryStatistics,
-						QueryStatistics::getCacheHitCount
-				)
-						.tags( tags )
-						.tags( "result", "hit", "query", query )
-						.description( "Number of query cache hits" )
-						.register( meterRegistry );
-
-				FunctionCounter.builder(
-						"hibernate.query.cache.requests",
-						queryStatistics,
-						QueryStatistics::getCacheMissCount
-				)
-						.tags( tags )
-						.tags( "result", "miss", "query", query )
-						.description( "Number of query cache misses" )
-						.register( meterRegistry );
-
-				FunctionCounter.builder(
-						"hibernate.query.cache.puts",
-						queryStatistics,
-						QueryStatistics::getCachePutCount
-				)
-						.tags( tags )
-						.tags( "query", query )
-						.description( "Number of cache puts for a query" )
-						.register( meterRegistry );
-
-				FunctionTimer.builder(
-						"hibernate.query.execution.total",
-						queryStatistics,
-						QueryStatistics::getExecutionCount,
-						QueryStatistics::getExecutionTotalTime,
-						TimeUnit.MILLISECONDS
-				)
-						.tags( tags )
-						.tags( "query", query )
-						.description( "Query executions" )
-						.register( meterRegistry );
-
-				TimeGauge.builder(
-						"hibernate.query.execution.max",
-						queryStatistics,
-						TimeUnit.MILLISECONDS,
-						QueryStatistics::getExecutionMaxTime
-				)
-						.tags( tags )
-						.tags( "query", query )
-						.description( "Query maximum execution time" )
-						.register( meterRegistry );
-
-				TimeGauge.builder(
-						"hibernate.query.execution.min",
-						queryStatistics,
-						TimeUnit.MILLISECONDS,
-						QueryStatistics::getExecutionMinTime
-				)
-						.tags( tags )
-						.tags( "query", query )
-						.description( "Query minimum execution time" )
-						.register( meterRegistry );
-
-				FunctionCounter.builder(
-						"hibernate.query.execution.rows",
-						queryStatistics,
-						QueryStatistics::getExecutionRowCount
-				)
-						.tags( tags )
-						.tags( "query", query )
-						.description( "Number of rows processed for a query" )
-						.register( meterRegistry );
-			}
-		}
+		super( sessionFactory, sessionFactoryName, tags );
 	}
 }
